// pr412-my-charity-change-app-reactnative/src/view/components/uiKit/Box/types.ts
export interface BoxProps {
  m?: string | number;
  mt?: string | number;
  mb?: string | number;
  mr?: string | number;
  ml?: string | number;
  p?: string | number;
  pt?: string | number;
  pb?: string | number;
  pv?: string | number;
  ph?: string | number;
  pr?: string | number;
  pl?: string | number;
  width?: string | number;
  height?: string | number;
  mh?: string;
  flex?: number;
  align?: string;
  bg?: string;
  opacity?: string | number;
  border?: string;
  br?: string | number;
  row?: boolean;
  wrap?: boolean;
  center?: boolean;
  justify?: string;
  spaceBetween?: boolean;
  spaceAround?: boolean;
}